Use UGII_USER_DIR in the user scope first
CURA-4445

It requires admin's permissions to read newly-set system-wide
environment variables. To improve the environment variable handling,
this commit checks for UGII_USER_DIR set in the user scope first and
then fall back to the system environment varaibles. This way if
UGII_USER_DIR is updated in the user-scope, the plugin can pick it up
immediately.